refactor(tests): remove deprecated __tests__
