+++ src/features/navigation/RootNavigator/linking/getScreenPath.test.ts
@@ -1,4 +1,4 @@
-import { getScreenPath } from '../getScreenPath'
+import { getScreenPath } from './getScreenPath'
 
 describe('getScreenPath()', () => {
   it.each`
