+++ src/features/navigation/RootNavigator/linking/getStateFromPath.test.ts
