+++ src/ui/theme/typographyAttrs/getNoHeadingAttrs.native.test.ts
